% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatPMCMRPlusResults}
\alias{formatPMCMRPlusResults}
\title{Helper function to formats the results of PMCMRPlus}
\usage{
formatPMCMRPlusResults(res, y)
}
\arguments{
\item{res}{Result from a test from formatPMCMRPlusResults}

\item{y}{the y position of the bracket}
}
\value{
A tibble
}
\description{
Helper function to formats the results of PMCMRPlus
}
