% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stat.R
\docType{data}
\name{GeomStat}
\alias{GeomStat}
\title{GeomStat}
\format{
An object of class \code{GeomStat} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomStat
}
\description{
Geom layer for plotting statistics
}
\keyword{datasets}
