% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kwAllPairsDunnTest.R
\name{stat_kwAllPairsDunnTest}
\alias{stat_kwAllPairsDunnTest}
\title{stat_kwAllPairsDunnTest}
\usage{
stat_kwAllPairsDunnTest(
  mapping = NULL,
  data = NULL,
  geom = GeomStat,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  size = 10,
  hide.ns = TRUE,
  tick.length = 0.02,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{Usually no need to change this}

\item{geom}{Geom used to render the brackets (default: GeomStat)}

\item{position}{Default: identity}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Has no effect.}

\item{size}{Fontsize for the annotation}

\item{hide.ns}{Should p-values lower than 0.05 be removed? Default TRUE}

\item{tick.length}{Length of the ticks in the p-value brackets. Default 0.02}

\item{inherit.aes}{TRUE}

\item{...}{Further arguments passed on to the layer in params}
}
\value{
A ggplot layer
}
\description{
stat_kwAllPairsDunnTest
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, cty)) +
  geom_boxplot()  +
  stat_kwAllPairsDunnTest()
}
