% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kwAllPairsDunnTest.R
\docType{data}
\name{StatKwAllPairsDunnTest}
\alias{StatKwAllPairsDunnTest}
\title{StatKwAllPairsDunnTest}
\format{
An object of class \code{StatKwAllPairsDunnTest} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatKwAllPairsDunnTest
}
\description{
Stat layer for calculating statistics
Uses \code{StatKwAllPairsDunnTest} from \code{PMCMRplus}
}
\keyword{datasets}
